feat: add optional data dir to blueprint manager
