+++ blueprint-manager/Cargo.toml
@@ -6,7 +6,7 @@ publish = false
 
 [[bin]]
 name = "blueprint-manager"
-path = "src/cli.rs"
+path = "src/main.rs"
 
 [dependencies]
 sp-core = { workspace = true }
@@ -18,7 +18,7 @@ structopt = { workspace = true }
 tangle-subxt = { workspace = true }
 toml = { workspace = true }
 hex = { workspace = true }
-tokio = { workspace = true, features = ["process", "io-util", "signal"] }
+tokio = { workspace = true, features = ["process", "io-util", "signal", "macros"] }
 reqwest = { workspace = true }
 sha2 = { workspace = true }
 futures = { workspace = true }
